refactor(featrues): mvdir into FeatureProvider
