+++ packages/bezier-react/src/components/AlphaSmoothCornersBox/AlphaSmoothCornersBox.tsx
@@ -2,14 +2,16 @@ import React, { forwardRef } from 'react'
 
 import classNames from 'classnames'
 
-import { FeatureType } from '~/src/features'
 import {
   cssUrl,
   cssVar,
   px,
 } from '~/src/utils/style'
 
-import { useFeatureFlag } from '~/src/components/FeatureProvider'
+import {
+  FeatureType,
+  useFeatureFlag,
+} from '~/src/components/FeatureProvider'
 
 import { type AlphaSmoothCornersBoxProps } from './AlphaSmoothCornersBox.types'
