+++ packages/bezier-react/src/components/FeatureProvider/Feature/SmoothCornersFeature/smoothCornersScript.ts
