+++ packages/bezier-react/src/components/FeatureProvider/FeatureProvider.tsx
@@ -4,9 +4,10 @@ import React, {
   useState,
 } from 'react'
 
-import { FeatureType } from '~/src/features'
 import { createContext } from '~/src/utils/react'
 import { isEmpty } from '~/src/utils/type'
+
+import { FeatureType } from '~/src/components/FeatureProvider'
 
 import {
   type FeatureFlag,
