+++ packages/bezier-react/src/components/FeatureProvider/FeatureProvider.types.ts
@@ -1,7 +1,7 @@
 import {
   type Feature,
   type FeatureType,
-} from '~/src/features'
+} from '~/src/components/FeatureProvider'
 import { type ChildrenProps } from '~/src/types/props'
 
 export type FeatureFlag = Record<FeatureType, boolean>
