+++ packages/bezier-react/src/components/FeatureProvider/index.ts
@@ -6,3 +6,4 @@ export type {
   FeatureFlag,
   FeatureProviderProps,
 } from './FeatureProvider.types'
+export * from './Feature'
